requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
pydantic==2.5.3
httpx==0.26.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
